[FEATURE] Support Vite Plugin Pug
