+++ docs/item-1.html
@@ -42,6 +42,11 @@
                         <span class="kss-nav__ref">1.2</span><span class="kss-nav__name">Button - Submit</span>
                     </a>
                   </li>
+                  <li class="kss-search__item kss-nav__menu-item ">
+                    <a class="kss-search__link kss-nav__menu-link" href="section-1.html#kssref-1-3" data-singlesectiontrigger="kssref-1-3">
+                        <span class="kss-nav__ref">1.3</span><span class="kss-nav__name">Paragraph</span>
+                    </a>
+                  </li>
               </ul>
         </li>
         <li class="kss-search__item kss-nav__menu-item">
@@ -94,6 +99,12 @@
                   <span class="kss-nav__menu-link" data-singlesectiontrigger="kssref-1-2">
                     <span class="kss-nav__ref ">1.2</span
                     ><span class="kss-nav__name">Button - Submit</span>
+                  </span>
+                </li>
+                <li class="kss-nav__menu-item">
+                  <span class="kss-nav__menu-link" data-singlesectiontrigger="kssref-1-3">
+                    <span class="kss-nav__ref ">1.3</span
+                    ><span class="kss-nav__name">Paragraph</span>
                   </span>
                 </li>
               </ul>
