+++ docs/section-colors.html
@@ -42,6 +42,11 @@
                         <span class="kss-nav__ref">1.2</span><span class="kss-nav__name">Button - Submit</span>
                     </a>
                   </li>
+                  <li class="kss-search__item kss-nav__menu-item ">
+                    <a class="kss-search__link kss-nav__menu-link" href="section-1.html#kssref-1-3" data-singlesectiontrigger="kssref-1-3">
+                        <span class="kss-nav__ref">1.3</span><span class="kss-nav__name">Paragraph</span>
+                    </a>
+                  </li>
               </ul>
         </li>
         <li class="kss-search__item kss-nav__menu-item">
