+++ lib/modules/modifierInsertSection.js
@@ -3,11 +3,12 @@
 const fs = require('fs');
 const path = require('path');
 const pug = require('pug');
+const replaceVitePugTags = require('../helpers/replaceVitePugTag');
 
 // eslint-disable-next-line no-useless-escape
 const regexModifierLine = /<insert-markup>(([0-9\.]*)\-?([0-9]*))<\/insert-markup>/m;
 
-module.exports = (Handlebars) => {
+module.exports = (Handlebars, options) => {
 	Handlebars.registerHelper('modifierInsertSection', (markup, sections, styleGuide, sourcePath) => {
 		let markupNumber;
 
@@ -102,17 +103,15 @@ module.exports = (Handlebars) => {
 			return markup;
 		};
 
-		/**
-		 * Replace <
-		 */
+		let output = replaceVitePugTags(markup, options);
 
 		// Init function
-		const insertLines = markup.match(regexModifierLine);
+		const insertLines = output.match(regexModifierLine);
 		// Check isset '<insert-markup>'
 		if (!insertLines) {
-			return markup;
+			return output;
 		}
 
-		return replaceInsert(markup, styleGuide);
+		return replaceInsert(output, styleGuide);
 	});
 };
