+++ source/scss/01_base/_paragraph.scss
@@ -1,5 +1,7 @@
 /*
 Paragraph
+
 Markup: <insert-vite-pug src="source/templates/02_components/paragraph.pug"></insert-vite-pug>
+
 Styleguide 1.3
 */
